requirements.txt
streamlit
plotly
pandas
psycopg2-binary
python-dotenv
requests
